Fix color sorter component template override
The Vue template compiler compiles the template of a .vue file
into a render function. The function has precedence over a template
declaration so the template of the color sorter component was ignored.
Now the component also compiles the template into a render function
and the original render function is correctly overridden.

Resolves #23